scaffold remaining decryptors
